refactor: move GameQuestion and AnswerResult to domain module
