+++ domain/src/main/java/com/amsterdam/domain/useCase/game/character/GenerateCharacterQuestionsUseCase.kt
@@ -4,8 +4,8 @@ package com.amsterdam.domain.useCase.game.character
 
 import com.amsterdam.domain.repository.GameRepository
 import com.amsterdam.domain.useCase.game.GetGameDifficultyByDifficultyTypeUseCase
+import com.amsterdam.domain.utils.GameQuestion
 import com.amsterdam.entity.GameDifficulty.DifficultyType
-import com.amsterdam.entity.GameQuestion
 import kotlin.uuid.ExperimentalUuidApi
 
 class GenerateCharacterQuestionsUseCase(
