+++ domain/src/main/java/com/amsterdam/domain/useCase/game/guessByPoster/GenerateMoviePosterQuestionUseCase.kt
@@ -2,8 +2,8 @@ package com.amsterdam.domain.useCase.game.guessByPoster
 
 import com.amsterdam.domain.repository.GameRepository
 import com.amsterdam.domain.useCase.game.GetGameDifficultyByDifficultyTypeUseCase
+import com.amsterdam.domain.utils.GameQuestion
 import com.amsterdam.entity.GameDifficulty
-import com.amsterdam.entity.GameQuestion
 
 
 class GenerateMoviePosterQuestionsUseCase(
