+++ domain/src/main/java/com/amsterdam/domain/useCase/game/releaseYear/DoGuessReleaseGameHintUseCase.kt
@@ -3,7 +3,7 @@ package com.amsterdam.domain.useCase.game.releaseYear
 import com.amsterdam.domain.exceptions.NotEnoughPointsException
 import com.amsterdam.domain.useCase.common.GetTotalUserPointsUseCase
 import com.amsterdam.domain.useCase.game.UpdateUserGamePointsUseCase
-import com.amsterdam.entity.GameQuestion
+import com.amsterdam.domain.utils.GameQuestion
 import kotlinx.coroutines.flow.first
 
 class DoGuessReleaseGameHintUseCase(
