+++ domain/src/main/java/com/amsterdam/domain/useCase/game/whichGenre/DoGuessGenreGameHintUseCase.kt
@@ -3,7 +3,7 @@ package com.amsterdam.domain.useCase.game.whichGenre
 import com.amsterdam.domain.exceptions.NotEnoughPointsException
 import com.amsterdam.domain.useCase.common.GetTotalUserPointsUseCase
 import com.amsterdam.domain.useCase.game.UpdateUserGamePointsUseCase
-import com.amsterdam.entity.GameQuestion
+import com.amsterdam.domain.utils.GameQuestion
 import com.amsterdam.entity.category.MovieGenre
 import kotlinx.coroutines.flow.first
