+++ domain/src/test/java/com/amsterdam/domain/useCase/game/character/DoGuessCharacterGameHintUseCaseTest.kt
@@ -3,7 +3,7 @@ package com.amsterdam.domain.useCase.game.character
 import com.amsterdam.domain.exceptions.NotEnoughPointsException
 import com.amsterdam.domain.useCase.common.GetTotalUserPointsUseCase
 import com.amsterdam.domain.useCase.game.UpdateUserGamePointsUseCase
-import com.amsterdam.entity.GameQuestion
+import com.amsterdam.domain.utils.GameQuestion
 import io.mockk.Runs
 import io.mockk.coEvery
 import io.mockk.just
