+++ domain/src/test/java/com/amsterdam/domain/useCase/game/whichGenre/DoGuessGenreGameHintUseCaseTest.kt
@@ -3,7 +3,7 @@ package com.amsterdam.domain.useCase.game.whichGenre
 import com.amsterdam.domain.exceptions.NotEnoughPointsException
 import com.amsterdam.domain.useCase.common.GetTotalUserPointsUseCase
 import com.amsterdam.domain.useCase.game.UpdateUserGamePointsUseCase
-import com.amsterdam.entity.GameQuestion
+import com.amsterdam.domain.utils.GameQuestion
 import com.amsterdam.entity.category.MovieGenre
 import com.google.common.truth.Truth.assertThat
 import io.mockk.Runs
